pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "event-harvester"
version = "0.1.0"
description = "Harvest Discord + Telegram messages, extract tasks via OpenRouter, create in TickTick"
requires-python = ">=3.10"
dependencies = [
    "telethon>=1.42",
    "litellm>=1.82.4,<2",
    "numpy>=1.24",
    "python-dotenv>=1.0",
    "ticktick-py>=2.0",
    "ccl-chromium-reader @ git+https://github.com/cclgroupltd/ccl_chromium_reader.git", "google-api-python-client>=2.193.0,<3", "google-auth-httplib2>=0.3.0,<0.4", "google-auth-oauthlib>=1.3.0,<2", "python-dateutil>=2.9.0.post0,<3",
    "scikit-learn>=1.3",
    "joblib>=1.3", "sentence-transformers>=3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "ruff>=0.4",
]

[project.scripts]
event-harvester = "event_harvester.cli:main_sync"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.pixi.workspace]
channels = ["conda-forge"]
platforms = ["win-64"]

[tool.pixi.dependencies]
python = ">=3.12,<3.13"

[tool.pixi.pypi-dependencies]
event-harvester = { path = ".", editable = true }

[tool.pixi.environments]
default = { solve-group = "default" }
dev = { features = ["dev"], solve-group = "default" }

[tool.pixi.tasks]

[tool.pixi.feature.dev.tasks]
test = "pytest -v"
lint = "ruff check src/ tests/"
format = "ruff format src/ tests/"
